+++ app/src/main/java/campus/tech/kakao/map/data/remote/dto/DocumentEntity.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.data.model
+package campus.tech.kakao.map.data.remote.dto
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
